more tests

+++ Data/Repo/ProductRepository.cpp
@@ -119,7 +119,7 @@ namespace productrepo {
         return false;
     }
 
-    bool ProductRepository::validate_stock(int id, int amount)
+    bool ProductRepository::validate_stock(int id, float amount)
     {
         if (amount > 0)
         {
